cebu: Fix touchscreen loading for R
Even without CONFIG_MODULE_SIG, touch doesn't work.
focaltech_0flash_mmi requires to be loaded in separately, before the
firmware.bin file can be flashed.

This is likely to be a temporary fix, as TWRP is gaining the feature
to load in kernel modules natively with a flag in BoardConfig.mk.

Change-Id: I16838db625cfdf8479d8f279b216cabc72ffbeee