cebu: Move to inbuilt TWRP logic for kernel module loading
TWRP recently gained code to insmod kernel modules natively. This commit
switches to that code, instead of using a hacky script to do the same.

With the notable exception of focaltech_0flash_mmi, all other modules
are using the new logic. Focaltech_0flash_mmi is very finicky, and
requires to be inserted at a specific point in the boot process to
work properly, otherwise touch doesn't work. It doesn't work with the
new logic, but works right now, inserting it in init.rc, so keeping
things as they are, for focaltech_0flash_mmi.

Change-Id: I0d61eed2e6d9f323dcc1aea6b77056f9a38cb0e9